Packages in domain refactor
