+++ src/main/java/com/alkemy/ong/domain/models/Role.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.domain.model;
+package com.alkemy.ong.domain.models;
 
 import java.time.LocalDateTime;
