+++ src/main/java/com/alkemy/ong/domain/services/NewsService.java
@@ -1,5 +1,5 @@
-package com.alkemy.ong.domain.service;
-import com.alkemy.ong.domain.model.News;
+package com.alkemy.ong.domain.services;
+import com.alkemy.ong.domain.models.News;
 import org.springframework.stereotype.Service;
 
 @Service
